requirements.txt
flask
moviepy
openai
Pillow
requests
python-dotenv
uuid
rq
google-auth-oauthlib
google-api-python-client
redis